🔨 Resource refactor 🔨
