+++ backend_v3/src/services/resourceServices.ts
@@ -1,4 +1,4 @@
-import * as resource from '../controllers/resource.controller'
+import * as resource from '../database/resource'
 import { Resource } from '../types'
 import { log } from '../utils/functions'
